tourism-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-mail==0.9.1
flask-caching==2.1.0
flask-limiter==3.5.0
pymysql==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
jieba==0.42.1
requests==2.31.0
pypinyin==0.49.0
beautifulsoup4==4.12.3
redis==5.0.1
lxml==5.1.0
Pillow==10.2.0
fake-useragent==1.4.0
numpy==1.24.3
scikit-learn==1.3.0 